requirements.txt
opencv-python
pillow
HumanCursor
pyautogui
pyperclip
pynput